Refactor calc_fig in _kaleido_tab.
