+++ src/py/kaleido/_kaleido_tab/_js_logger.py
@@ -1,0 +1,45 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import logistro
+
+_logger = logistro.getLogger(__name__)
+
+if TYPE_CHECKING:
+    from typing import Any
+
+    import choreographer
+
+
+def _make_console_logger(name, log):  # TODO(AJP) how do we use this
+    """Create printer specifically for console events. Helper function."""
+
+    async def console_printer(event):
+        _logger.debug2(f"{name}:{event}")  # TODO(AJP): parse? stretch goal.
+        log.append(str(event))
+
+    return console_printer
+
+
+class JavascriptLogger:
+    log: list[Any]
+    """A list of console outputs from the tab."""
+
+    def __init__(self) -> None:
+        self.log = []
+
+    def activate(self, tab: choreographer.Tab):
+        tab.unsubscribe("Runtime.consoleAPICalled")
+        tab.subscribe(
+            "Runtime.consoleAPICalled",
+            _make_console_logger("tab js console", self.javascript_log),
+        )
+
+    def reset(self, tab: choreographer.Tab):
+        tab.unsubscribe("Runtime.consoleAPICalled")
+        self.log = []
+        tab.subscribe(
+            "Runtime.consoleAPICalled",
+            _make_console_logger("tab js console", self.javascript_log),
+        )
